Update instructions
I always hated this part of the instructions and I've wanted to change it from the beginning.

I think I wrote it this way because I didn't know how environment variables worked with Wine. Apparently, it just directly provides the host system's environment variables to the application it runs, so specifying the OS and whether Wine is being used is mostly irrelevant.

The user may need to adapt the actual variable value on Linux/Unix but it's probably fine to leave that out - if anyone has trouble with the variable, they'll probably naturally try copying the license file as described anyway.